Ready for migration (without notifications)

+++ db/migrate/20131214181854_applications_to_postulations.rb
@@ -1,5 +1,3 @@
-require 'open-uri'
-
 class ApplicationsToPostulations < ActiveRecord::Migration
   def up
     Application.all.each do |application|
@@ -9,20 +7,21 @@ class ApplicationsToPostulations < ActiveRecord::Migration
         user.fullname = application.name
         user.university_id = application.university_id
         user.degree = application.degree
-        io = open(URI.parse(application.resume.url))
-        user.resume = io
         user.save(validate: false)
       end
       Postulation.create user: user, advertisement: application.advertisement
+      puts "Application to postulation (#{user.email})"
     end
   end
 
   def down
     Application.all.each do |application|
       u = User.where(email: application.email).first
-      p = Postulation.where(user: u).first
-      u.destroy
-      p.destroy
+      if u
+        p = Postulation.where(user_id: u.id).first
+        u.destroy
+        p.destroy
+      end
     end    
   end
 end
